+++ test/preview/components/generators/openai/test_gpt35_generator.py
@@ -1,3 +1,4 @@
+import os
 from unittest.mock import patch, Mock
 from copy import deepcopy
 
@@ -263,3 +264,56 @@ class TestGPT35Generator:
             "2 out of the 4 completions have been truncated before reaching a natural "
             "stopping point. Increase the max_tokens parameter to allow for longer completions."
         )
+
+    @pytest.mark.skipif(
+        not os.environ.get("OPENAI_API_KEY", None),
+        reason="Export an env var called OPENAI_API_KEY containing the OpenAI API key to run this test.",
+    )
+    @pytest.mark.integration
+    def test_gpt35_generator_run():
+        component = GPT35Generator(api_key=os.environ.get("OPENAI_API_KEY"), n=1)
+        results = component.run(prompt="What's the capital of France?")
+        assert len(results["replies"]) == 1
+        assert "Paris" in results["replies"][0]
+        assert len(results["metadata"]) == 1
+        assert "gpt-3.5" in results["metadata"][0]["model"]
+        assert results["metadata"][0]["finish_reason"] == "stop"
+
+    @pytest.mark.skipif(
+        not os.environ.get("OPENAI_API_KEY", None),
+        reason="Export an env var called OPENAI_API_KEY containing the OpenAI API key to run this test.",
+    )
+    @pytest.mark.integration
+    def test_gpt35_generator_run_wrong_model_name():
+        component = GPT35Generator(
+            model_name="something-obviously-wrong", api_key=os.environ.get("OPENAI_API_KEY"), n=1
+        )
+        with pytest.raises(openai.InvalidRequestError, match="The model `something-obviously-wrong` does not exist"):
+            component.run(prompt="What's the capital of France?")
+
+    @pytest.mark.skipif(
+        not os.environ.get("OPENAI_API_KEY", None),
+        reason="Export an env var called OPENAI_API_KEY containing the OpenAI API key to run this test.",
+    )
+    @pytest.mark.integration
+    def test_gpt35_generator_run_streaming():
+        class Callback:
+            def __init__(self):
+                self.responses = ""
+
+            def __call__(self, chunk):
+                self.responses += chunk.choices[0].delta.content if chunk.choices[0].delta else ""
+                return chunk
+
+        callback = Callback()
+        component = GPT35Generator(os.environ.get("OPENAI_API_KEY"), streaming_callback=callback, n=1)
+        results = component.run(prompt="What's the capital of France?")
+
+        assert len(results["replies"]) == 1
+        assert "Paris" in results["replies"][0]
+
+        assert len(results["metadata"]) == 1
+        assert "gpt-3.5" in results["metadata"][0]["model"]
+        assert results["metadata"][0]["finish_reason"] == "stop"
+
+        assert callback.responses == results["replies"][0]
